requirements.txt
azure-cli
azdev
